Added El Jardin Portfolio and Delete User CRUD, advancing
