+++ src/interfaces/backend/user.ts
@@ -2,3 +2,10 @@ export interface UserLogin {
   email: string;
   password: string;
 }
+
+export interface UserDelete {
+  _id: string;
+  id: string;
+  name?: string;
+  email?: string;
+}
